Move all code from Cardano.Api.LedgerState to Cardano.Api.LedgerState.Core
