+++ cardano-api/cardano-api.cabal
@@ -106,6 +106,7 @@ library internal
                         Cardano.Api.Keys.Shelley
                         Cardano.Api.Ledger.Lens
                         Cardano.Api.LedgerState
+                        Cardano.Api.LedgerState.Core
                         Cardano.Api.LedgerEvents.ConvertLedgerEvent
                         Cardano.Api.LedgerEvents.Rule.BBODY.DELEGS
                         Cardano.Api.LedgerEvents.Rule.BBODY.LEDGER
